+++ tests/testing_data/config_fl_filters.json
@@ -1,12 +1,12 @@
 {
     "pre_filters": [
         {
-            "_target_": "monai.fl.utils.dummy_filter.DummyFilter"
+            "_target_": "monai.fl.utils.filters.SummaryFilter"
         }
     ],
     "post_weight_filters": [
         {
-            "_target_": "monai.fl.utils.dummy_filter.DummyFilter"
+            "_target_": "monai.fl.utils.filters.SummaryFilter"
         }
     ],
     "post_evaluate_filters": []
